# docs/swagger.yaml
basePath: /
definitions:
...
  models.Login:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  models.MovieFilter:
    properties:
      genres:
        description: hasil string_agg
        type: string
      id:
        type: integer
      poster:
        type: string
      title:
        type: string
    type: object
  models.MovieResponse:
    properties:
      background_poster:
        type: string
      created_at:
        type: string
      director_id:
        type: integer
      duration:
        type: string
      id:
        type: integer
      popularity:
        type: integer
      poster:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.OrderHistory:
    properties:
      cinema_name:
        type: string
      created_at:
        type: string
      date:
        type: string
      is_paid:
        type: boolean
      location:
        type: string
      movie_title:
        type: string
      order_id:
        type: integer
      qr_code:
        type: string
      seats:
        type: string
      time:
        type: string
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Register:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
        x-omitempty: true
    required:
    - email
    - password
    type: object
  models.Response:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ticketing
  title: tickitz
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan email dan password
      parameters:
      - description: request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login user
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun baru
      parameters:
      - description: Register request (role optional, default user)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register user baru
      tags:
      - auth
  /movie/:
    get:
      description: Get movies by title and/or genre with pagination
      parameters:
      - description: Filter by title
        in: query
        name: title
        type: string
      - description: Filter by genre
        in: query
        name: genre
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieFilter'
            type: array
      summary: Get Movies by Filter
      tags:
      - movies
  /movie/popular:
    get:
      description: Get list of popular movies ordered by popularity
      parameters:
      - description: Limit number of movies (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieResponse'
            type: array
      summary: Get Popular Movies
      tags:
      - movies
...
  /movies/{movie_id}:
    get:
      consumes:
      - application/json
      description: Get movie detail by ID
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie detail
      tags:
      - movies
...
  /user/history:
    get:
      description: Get all order history for a user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrderHistory'
            type: array
      security:
      - JWTtoken: []
      summary: Get order history
      tags:
      - history
...
securityDefinitions:
  JWTtoken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"